refactor: test 코드 파일 구조 변경
