+++ README.md
@@ -36,19 +36,32 @@
 
 # 테스트 목록
 
+## domain
+
 1. Car
 
-- 경주할 자동차 이름 입력 검증
-  - [x] 자동차는 이름을 가져야한다.
-  - [x] 자동차 이름은 1~5자여야 한다.
-  - [x] 자동차는 쉼표를 기준으로 구분한다.
-  - [x] 자동차는 2대 이상이어야 한다.
-  - [x] 자동차 이름은 중복되면 안된다.
-- 자동차가 잘 움직이는지 검증
-  - [x] 자동차는 4 이상의 숫자를 받으면 전진할 수 있다.
+- [x] 자동차는 이름을 가져야한다.
+- [ ] 자동차는 움직일 수 있다.
 
 2. Race
 
-- 경주 시도 횟수 입력 검증
-  - [x] 시도 횟수는 1~20 사이여야 한다.
-  - [x] 시도 횟수는 자연수로 입력해야 한다.
+- [ ] 자동차는 4 이상의 숫자를 받으면 전진할 수 있다.
+- [ ] 가장 멀리 떨어진 위치에 있는 자동차가 우승자다.
+
+3. CarPositionHistory
+
+- [ ] 각 자동차가 움직인 기록을 가지고 있다.
+
+## validation
+
+1. carValidates
+
+- [x] 자동차 이름은 1~5자여야 한다.
+- [x] 자동차는 쉼표를 기준으로 구분한다.
+- [x] 자동차는 2대 이상이어야 한다.
+- [x] 자동차 이름은 중복되면 안된다.
+
+2. tryCountValidates
+
+- [x] 시도 횟수는 1~20 사이여야 한다.
+- [x] 시도 횟수는 자연수로 입력해야 한다.
