+++ __tests__/Race.test.js
@@ -1,16 +1,11 @@
-import { checkIsInteger, checkTryCountRange } from '../src/validation/tryCountValidates';
+import Car from '../src/domain/Car';
+import { CAR_MOVE_STANDARD } from '../src/constants/MAGIC_NUMBER';
 
-describe('경주 시도 횟수 입력 검증 테스트', () => {
-  test.each([1, 20, 5])('시도 횟수는 1~20 사이여야 한다.', (tryCount) => {
-    expect(() => {
-      checkTryCountRange(tryCount);
-    }).not.toThrow('[ERROR]');
-  });
+describe('자동차가 잘 움직이는지 테스트', () => {
+  const car = new Car('상추');
 
-  test('시도 횟수는 자연수로 입력해야 한다.', () => {
-    const tryCount = 1.2;
-    expect(() => {
-      checkIsInteger(tryCount);
-    }).toThrow('[ERROR]');
+  car.move(4 >= CAR_MOVE_STANDARD);
+  test('자동차는 4 이상의 숫자를 받으면 전진할 수 있다.', () => {
+    expect(car.position).toBe(1);
   });
 });
